requirements.txt
demosys-py==0.2.1
